Added time-wise functionality for drones, as well as updating order status from both client and server side

+++ src/main/java/comp1206/sushi/common/Basket.java
@@ -23,7 +23,7 @@ public class Basket implements Serializable{
 	
 	public void addDishToBasket(Dish dish, Number quantity) {
 		if (basketMap.containsKey(dish)) {
-			cost -= (Integer) dish.getPrice() * (Integer) basketMap.get(dish);
+			cost -= (Float) dish.getPrice() * (Integer) basketMap.get(dish);
 		}
 		cost += ((Float) dish.getPrice() * (Integer) quantity);
 		basketMap.put(dish, quantity);
